Jetson dev

+++ Main/Py_Modules/Stereo_Camera.py
@@ -41,6 +41,7 @@ class Stereo_Camera:
         
         #---------        
         #Start up Depth Camera; also boots Ros subscribers
+        self.CAMprocess=None; self.SpinThread=None #prevent minor error in destructor
         self.Disparity_sub=None; self.ColorImg_sub=None
         if self.Real:
             #Start up Depth Camera; also boots Ros subscribers
@@ -64,18 +65,19 @@ class Stereo_Camera:
     def __del__(self):
         #-------------
         #Parallel Terminal
-        prALERT(f'Stereo_Camera Destructor:\tKilling Camera Startup Parallel Terminal\n{"="*12}')
-        try:
-            prYellow("Giving time for Split terminal to run before closing")
-            time.sleep(5)
-            #os.kill(self.CAMprocess.pid,signal.SIGTERM)
-            os.system("pkill -f MS_startup.sh")
-            if self.CAMprocess.poll() is not None: prRed("Couldn't shutdown parallel terminal; please shutdown popped up terminal yourself")
-        except Exception as e:
-            print("error shutting down parallel terminal:",e)
+        if not self.CAMprocess is None:
+            prALERT(f'Stereo_Camera Destructor:\tKilling Camera Startup Parallel Terminal\n{"="*12}')
+            try:
+                prYellow("Giving time for Split terminal to run before closing")
+                time.sleep(5)
+                #os.kill(self.CAMprocess.pid,signal.SIGTERM)
+                os.system("pkill -f MS_startup.sh")
+                if self.CAMprocess.poll() is not None: prRed("Couldn't shutdown parallel terminal; please shutdown popped up terminal yourself")
+            except Exception as e:
+                prRed("error shutting down parallel terminal,Likely minor:\n",e)
         #-------------
         #Kill threads
-        if self.multithread:
+        if self.multithread and not self.SpinThread is None:
             prALERT(f'Stereo_Camera Destructor:\tKilling Parallel Node Spin Threads\n{"="*12}')  
             try:
                 #self.t1.raise_exception()
@@ -343,16 +345,18 @@ def balance_numpy(arr):
 
 
 if __name__ == "__main__":
-    cammie = Stereo_Camera()
+    try:
+        cammie = Stereo_Camera()
     
-    print("waiting (safe)...")
-    time.sleep(4)
-    print(f"wait done\n\n{'-'*24}\n")
+        print("waiting (safe)...")
+        time.sleep(4)
+        print(f"wait done\n\n{'-'*24}\n")
     
-    while True:
-        cv2.imshow("Depthmap <q key to quit>",balance_numpy(cammie.Depth_Map))
-        cv2.imshow("CameraFeed <q key to quit>",cammie.get_feed())
-        if cv2.waitKey(1) == ord('q'): break    
-    os.system("pkill -f MS_startup.sh")
+        while True:
+            cv2.imshow("Depthmap <q key to quit>",balance_numpy(cammie.Depth_Map))
+            cv2.imshow("CameraFeed <q key to quit>",cammie.get_feed())
+            if cv2.waitKey(1) == ord('q'): break    
+    except:
+        os.system("pkill -f MS_startup.sh")
